+++ app/assets/stylesheets/dev_toolbar.css
@@ -14,6 +14,7 @@
 }
 
 #dev-toolbar-toggle {
+  all: unset;
   font-size: 2em;
   border: none;
   cursor: pointer;
