--- lib/dev_toolbar/app/controllers/dev_toolbar/erd_controller.rb
@@ -1,11 +1,0 @@
-module DevToolbar
-  class ErdController < ActionController::Base
-    layout false
-    
-    def show
-      @erd_path = Rails.root.join("erd.png")
-      # render :show
-      render :show
-    end
-  end
-end
